feat(core): add `findMarkRange`
