+++ .changeset/find-mark-range.md
@@ -1,0 +1,6 @@
+---
+"@prosekit/core": minor
+"prosekit": minor
+---
+
+Add `findMarkRange`, which returns the contiguous range of the first mark matching a predicate that touches a resolved position.
