+++ packages/core/src/index.ts
@@ -125,10 +125,11 @@ export { clsx } from './utils/clsx.ts'
 export { containsInlineNode } from './utils/contains-inline-node.ts'
 export { defaultBlockAt } from './utils/default-block-at.ts'
 export { isApple } from './utils/env.ts'
+export { findMarkRange, type MarkRange } from './utils/find-mark-range.ts'
 export { findNode, findNodes, type FindNodeResult } from './utils/find-node.ts'
 export { findParentNodeOfType } from './utils/find-parent-node-of-type.ts'
 export { findParentNode, type FindParentNodeResult } from './utils/find-parent-node.ts'
-export { getMarkRange, type MarkRange } from './utils/get-mark-range.ts'
+export { getMarkRange } from './utils/get-mark-range.ts'
 export { getMarkType } from './utils/get-mark-type.ts'
 export { getNodeType } from './utils/get-node-type.ts'
 export { isAtBlockStart } from './utils/is-at-block-start.ts'
